Only output the handled pipelines in the metadata of Out.
The metadata of Out now only loops over the pipelines
defined in the Out parameters instead of looping over
all the pipelines of the given team, reducing the
clutter displayed on the job.

The test case for the Pipelines fly command is also
removed from the Out test as it's no longer called.

Signed-off-by: Alessandro Degano <alessandro.degano@pix4d.com>